Refactorings, more tests
