+++ UpdateDB/DataLoading/Gems/IGemReader.cs
@@ -1,6 +1,6 @@
-﻿using System.Net.Http;
+﻿using POESKillTree.SkillTreeFiles;
+using System.Net.Http;
 using System.Threading.Tasks;
-using Gem = POESKillTree.SkillTreeFiles.ItemDB.Gem;
 
 namespace UpdateDB.DataLoading.Gems
 {
