+++ WPFSKillTree/SkillTreeFiles/SkillTree-Drawing.cs
@@ -61,6 +61,10 @@ namespace POESKillTree.SkillTreeFiles
         private DrawingVisual _ascActiveSkillIcons;
         private DrawingVisual _ascNodeSurround;
         private DrawingVisual _ascActiveNodeSurround;
+
+        public SkillTree()
+        {
+        }
         #endregion
         public void InitialSkillTreeDrawing()
         {
